Create Simulator which passes voltage to root bit
The idea is to have a simulator which on a tick interval will update
the voltage and pass it to it's root Bit; the bit then must update
itself and pass the voltage to its children, and so on and so forth.

This recursive behaviour is simple to implement and follows a simple
version of the interpreter pattern.  There's much work to be done
still:

- Create a (higher-order?) component which encapsulates the "frame" of
  a Little Bit
  - has a prop for the frame's color
  - defines where child bits will be placed
  - Encapsulates the drag-and-drop behaviour
- Define a widget/widget component for each little bit which defines
  what goes on in the center of the bit (having a speaker or a dial,
  etc)