Minimal demo including 2 different Little Bits
At this point, it is ready to add drag-and-drop functionality.

Frames can now show a label.  however, because using the z-index to
hide the interlock blocked input events on the Dimmer's slider.

The Bargraph currently does not use the label feature, but I might
just place a voltage readout there on hover.