Re-create application with yarn
create-react-app was causing all sorts of errors on import of
react-dnd; to circumvent the errors I re-generated the project
with yarn and moved the .git repository and appropriate files from
the old version.

I don't know why this fixed the errors, but it did.  Webpack was
failing to properly include the new dependencies, perhaps
create-react-app is not intended for use with yarn?