Create simple AST in Simulator
Create a recursive object which descirbes the tree of little bits
by deconstructing and re-constructing props.  This allows a change
in state in the simulator to re-render the entire tree, but also
keeps the ergonomics of being able to simply represent the tree
as nested React Elements when testing or creating an initial
layout!