Replace Frame drop target with special AST node
Create a placeholder node which itself has the drop target hook;
by passing the callback into the renderAST function, it is now
possible to register a callback to the placeholder node from within
the Simulator's constructor that is called on Drop; this means all
the information needed to start re-writing AST's is available in
one spot.

Also allows the frame to not care about it's children anymore.